+++ CHANGELOG.md
@@ -7,7 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [3.0.31] - 2025-10-07
+### Fixed
+
+- Some basic refacto: mutualized code, style
+- RainSpeed transpose added + fixed
+- Cartesian earth usage enlarged
+
+## [3.0.32] - 2025-10-07
 
 ### Fixed